Rename editorState -> editor
