+++ src/store/modules/index.js
@@ -6,18 +6,18 @@ import type { Location } from 'react-router'
 
 import entities from './entities'
 import exportAsPdf from './exportAsPdf'
-import editorState from './editorState'
+import editor from './editor'
 import preview from './preview'
 
 import type { EntitiesState } from './entities'
 import type { ExportAsPdfState } from './exportAsPdf'
-import type { EditorState } from './editorState'
+import type { EditorState } from './editor'
 import type { PreviewState } from './preview'
 
 export type RootState = {
   entities: EntitiesState,
   exportAsPdf: ExportAsPdfState,
-  editorState: EditorState,
+  editor: EditorState,
   preview: PreviewState,
   router: ?Location,
 }
@@ -25,7 +25,7 @@ export type RootState = {
 export default combineReducers({
   entities,
   exportAsPdf,
-  editorState,
+  editor,
   preview,
   router: routerReducer,
 })
