+++ src/store/selectors/pages.js
@@ -4,7 +4,7 @@ import { createSelector } from 'reselect'
 import type Page from 'entities/Page'
 import type { Position } from 'types'
 
-import { getCursorPosition } from './editorState'
+import { getCursorPosition } from './editor'
 import { createGetPageByCursorPosition } from './entities/pages'
 
 export const getCurrentPage = createSelector(
